requirements.txt
streamlit==1.50.0
requests==2.32.5
google-generativeai==0.8.5
python-dotenv==1.0.0
openai==1.0.0